+++ navable-api/src/main/java/com/unicamp/navable_api/api/model/OcorrenciaDTO.java
@@ -10,7 +10,7 @@ public class OcorrenciaDTO {
     private Integer idUsuario;
     private Integer idTipoOcorrencia;
     private TipoOcorrenciaDTO tipoOcorrencia;
-    private Integer latitude;
-    private Integer longitude;
+    private Double latitude;
+    private Double longitude;
 }
