+++ navable-api/src/main/java/com/unicamp/navable_api/persistance/entities/Estabelecimento.java
@@ -14,8 +14,8 @@ public class Estabelecimento {
     private Integer idEstabelecimento;
     private Integer idTipoEstabeleci;
     private String nome;
-    private Integer latitude;
-    private Integer longitude;
+    private Double latitude;
+    private Double longitude;
     private String imagem;
     private String endereco;
 }
